Added existance checks to imported values.
Locally we keep security settings in secret_settings.py. But on Heroku
those variables should not be loaded at all in settings.py. A fast
solution that will work for now is to have an empty secret_settings on
Heroku and use hasattr() in settings.py to only set values if they exist.