Updated Message date fields
These are the "created" and "updated" fields. I replaced the auto_now
functionality default values and a custom save() method, since auto_now
and auto_now_add are considered problematic.
I also changed the field names to have a date_ prefix, to be consistent
with django built-in name conventions, and because it makes it easier
to see at first glance what they are (just "created" could be a boolean,
date or something else.)
I also changed the admin view to still not be able to edit the created
and updated fields, since this would now sudden;y be possible using with
the new implementation.